rx tx mavlink: no, seq cannot be used to detect frame losses, because of routing with targets

+++ mLRS/Common/fifo.h
@@ -13,7 +13,7 @@
 
 #include <inttypes.h>
 
-template <class T, int FIFO_SIZE>
+template <class T, uint16_t FIFO_SIZE>
 class FifoBase
 {
   public:
